CLI: rename instances group -> params instantiate; drop remote-CRUD verbs
"Instance" is the backend, tied-to-a-service concept; the local kind is just
parameters. So the CLI group becomes `usvc_seller params` and the create action
becomes `params instantiate` (the params-kind analog of `specs upload`).

- `usvc_seller instances create`  -> `usvc_seller params instantiate`
- drop `instances list/show/delete` (remote-CRUD) — a service already carries
  its template provenance (managed_by_template), so use `services list`; remote
  instance management stays in the SDK (`client.instances`).
- `client.templates` / `client.instances` SDK unchanged (they wrap the backend).
- docs + cli-reference updated.

See unitysvc#1263 for the full local-layout/naming design.
