+++ src/unitysvc_sellers/cli.py
@@ -8,7 +8,7 @@ import typer
 
 from . import data
 from .commands import groups as groups_cmd
-from .commands import instances as instances_cmd
+from .commands import params as params_cmd
 from .commands import promotions as promotions_cmd
 from .commands import secrets as secrets_cmd
 from .commands import services as services_cmd
@@ -59,7 +59,7 @@ app.add_typer(promotions_cmd.app, name="promotions")
 app.add_typer(groups_cmd.app, name="groups")
 app.add_typer(secrets_cmd.app, name="secrets")
 app.add_typer(templates_cmd.app, name="templates")
-app.add_typer(instances_cmd.app, name="instances")
+app.add_typer(params_cmd.app, name="params")
 
 
 if __name__ == "__main__":
